Fix compiling with GCC 9+. Have to workaround a GCC bugfeature
See https://sourceforge.net/p/fbc/bugs/904/

Also fix or ignore a few GCC warnings. You should now see no warnings during
a compile! It seems GCC 9 fixes some previously spurious warnings.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@11170 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ surface.cpp
@@ -29,7 +29,7 @@ mutex surfaceMutex;
 
 // Print out all Surfaces, return the number
 int gfx_debugSurfaces_SW() {
-	debuginfo("%d Surfaces:", g_surfaces.size());
+	debuginfo("%d Surfaces:", (int)g_surfaces.size());
 	for (auto pSurf : g_surfaces) {
 		debuginfo("%p %d*%d refc=%d view=%d 32bit=%d base_surf=%p", pSurf, pSurf->width, pSurf->height, pSurf->refcount, pSurf->isview, (pSurf->format == SF_32bit), pSurf->base_surf);
 	}
